New leadlang importer

+++ leadman/src/packages/linker/mod.rs
@@ -13,7 +13,7 @@ mod script;
 
 pub enum ScriptClass {
   Pre,
-  Post
+  Post,
 }
 
 pub async fn run_script(meta: Arc<MetaPtr>, class: ScriptClass, print: MultiProgress) {
@@ -24,11 +24,18 @@ pub async fn run_script(meta: Arc<MetaPtr>, class: ScriptClass, print: MultiProg
 
     let cwd = format!("./.pkgcache/{hash}");
 
-    let meta: LibraryMeta = from_str(&fs::read_to_string(format!("{cwd}/pkgcache")).await.expect("Unable to read metadata")).expect("Invalid Metadata");
+    let meta: LibraryMeta = from_str(
+      &fs::read_to_string(format!("{cwd}/pkgcache"))
+        .await
+        .expect("Unable to read metadata"),
+    )
+    .expect("Invalid Metadata");
 
     let script = match class {
-      ScriptClass::Pre => meta.preinstall,
-      ScriptClass::Post => meta.postinstall
+      ScriptClass::Pre => {
+        meta.preinstall
+      },
+      ScriptClass::Post => meta.postinstall,
     };
 
     if let Some(script) = script {
@@ -43,44 +50,69 @@ pub async fn link(meta: Arc<MetaPtr>, print: MultiProgress) {
   let is0 = meta.pkver == 0;
 
   let _ = fs::remove_dir_all("./.lead_libs").await;
-  fs::create_dir_all("./.lead_libs").await.expect("Unable to rebuild links");
+  fs::create_dir_all("./.lead_libs")
+    .await
+    .expect("Unable to rebuild links");
 
   for (k, v) in &meta.dependencies {
     let hash = digest(format!("{k}@{v}"));
 
     let meta = format!("./.pkgcache/{hash}/pkgcache");
 
-    let resp: LibraryMeta = from_str(&fs::read_to_string(meta).await.expect("Error reading")).expect("Unable to parse");
+    let resp: LibraryMeta =
+      from_str(&fs::read_to_string(meta).await.expect("Error reading")).expect("Unable to parse");
 
     let mut platform_cwd = format!("./.pkgcache/{hash}/lib/{}", TARGET);
 
-    if !fs::metadata(&platform_cwd).await.expect("Unable to get metadata").is_dir() {
+    if !fs::metadata(&platform_cwd)
+      .await
+      .expect("Unable to get metadata")
+      .is_dir()
+    {
       platform_cwd = format!("./.pkgcache/{hash}/lib/{}", *OTHER_TARGET);
-      if !fs::metadata(&platform_cwd).await.expect("Unable to get metadata").is_dir() {
-        panic!("No Build for {k}@{v} is availble for {TARGET}");
+      if !fs::metadata(&platform_cwd)
+        .await
+        .expect("Unable to get metadata")
+        .is_dir()
+      {
+        panic!("No Build for {k}@{v} is availble for {TARGET}, run compile script (if available) for the crate from ./.pkgcache/{hash}");
       }
     }
 
-    copy_dir(&platform_cwd, format!("./.lead_libs/{hash}_{}", &resp.package)).await;
+    copy_dir(
+      &platform_cwd,
+      format!("./.lead_libs/{hash}_{}", &resp.package),
+    )
+    .await;
 
     let doc = format!("./.lead_libs/{hash}_{}/docs", &resp.package);
 
-    fs::write(format!("./.lead_libs/{hash}_{}/lead.lookup.lkp", &resp.package), resp.package).await;
+    fs::write(
+      format!("./.lead_libs/{hash}_{}/lead.lookup.lkp", &resp.package),
+      resp.package,
+    )
+    .await;
 
     if is0 {
       match fs::metadata(&doc).await {
-        Ok(m) => if m.is_file() {
-          print.suspend(|| {
-            println!("      ⚠️  Replacing file named docs in {k}@{v} for target {TARGET}");
-          });
-          fs::remove_file(&doc).await;
-          fs::copy(format!("./.pkgcache/{hash}/docs"), doc).await.expect("Unable to copy docs");
+        Ok(m) => {
+          if m.is_file() {
+            print.suspend(|| {
+              println!("      ⚠️  Replacing file named docs in {k}@{v} for target {TARGET}");
+            });
+            fs::remove_file(&doc).await;
+            fs::copy(format!("./.pkgcache/{hash}/docs"), doc)
+              .await
+              .expect("Unable to copy docs");
+          }
         }
         _ => {
           print.suspend(|| {
             println!("      ⚠️  No docs found for {k}@{v} for target {TARGET}, using generic docs");
           });
-          fs::copy(format!("./.pkgcache/{hash}/docs"), doc).await.expect("Unable to copy docs");
+          fs::copy(format!("./.pkgcache/{hash}/docs"), doc)
+            .await
+            .expect("Unable to copy docs");
         }
       }
     }
